backend/requirements.txt
# DealSnap - Python FastAPI Backend
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Data Validation
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Financial Calculations (using pure Python alternatives for better compatibility)
# numpy and scipy removed - calculator.py uses pure Python math

# OAuth/SSO Authentication
authlib>=1.3.0
itsdangerous>=2.1.0
httpx>=0.26.0
python-jose[cryptography]>=3.3.0

# CORS and HTTP
starlette>=0.36.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Development
python-dotenv>=1.0.0
